requirements.txt
matplotlib>=3.10.6
matplotlib-inline>=0.1.6
pandas>=2.3.2
pyarrow>=21.0.0
pybind11-stubgen>=2.5.5
thop>=0.1.1
timm>=1.0.19
PyYAML>=6.0
Pillow>=10.0.0
tqdm>=4.66.0
torchinfo>=1.8.0
